setup skeleton CreateEvent page (modularized into components)
